just drop numerics for now
