+++ model_templates/training/python3_pytorch/custom.py
@@ -2,9 +2,14 @@ from typing import List, Optional
 import pandas as pd
 import numpy as np
 
-import pickle
-
-from pipeline_utils import make_classifier, make_regressor
+from model_utils import (
+    build_regressor,
+    build_classifier,
+    train_regressor,
+    train_classifier,
+    save_torch_model,
+    subset_data,
+)
 
 
 def fit(
@@ -17,10 +22,8 @@ def fit(
 ) -> None:
     """
     This hook must be implemented with your fitting code, for running drum in the fit mode.
-
     This hook MUST ALWAYS be implemented for custom training models.
     For inference models, this hook can stick around unimplemented, and won’t be triggered.
-
     Parameters
     ----------
     X: pd.DataFrame - training data to perform fit on
@@ -38,19 +41,37 @@ def fit(
         custom models support this. There are two situations when values will be passed into
         row_weights, during smart downsampling and when weights are explicitly provided by the user
     kwargs: Added for forwards compatibility
-
     Returns
     -------
     Nothing
     """
-
+    # keep only numeric features
+    X_train = subset_data(X)
     # Feel free to delete which ever one of these you aren't using
     if class_order:
-        estimator = make_classifier(X)
+        estimator, optimizer, criterion = build_classifier(X_train)
+        train_classifier(X_train, y, estimator, optimizer, criterion)
+        artifact_name = "torch_bin.pth"
     else:
-        estimator = make_regressor(X)
+        estimator, optimizer, criterion = build_regressor(X_train)
+        train_regressor(X_train, y, estimator, optimizer, criterion)
+        artifact_name = "torch_reg.pth"
 
-    estimator.fit(X, y)
     # NOTE: We currently set a 10GB limit to the size of the serialized model
-    with open("{}/artifact.pkl".format(output_dir), "wb") as fp:
-        pickle.dump(estimator, fp)
+    save_torch_model(estimator, output_dir, artifact_name)
+
+
+def transform(data, model):
+    """
+    apply the same subsetting at prediction time as during fit
+
+    Parameters
+    ----------
+    data : is the dataframe given to **drum** to make predictions on
+    model : is the deserialized model loaded by **drum** or by `load_model`, if supplied
+
+    Returns
+    -------
+    Transformed data
+    """
+    return subset_data(data)
